pyproject.toml
[tool.poetry]
name = "submind"
version = "0.1.0"
description = ""
authors = ["Leo Guinan <leo.guinan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<3.13"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.29"
langchain = "^0.1.14"
langchain-openai = "^0.1.1"
pinecone-client = "^3.2.2"
python-decouple = "^3.8"
langchain-pinecone = "^0.0.3"
pymongo = "^4.6.3"
anthropic = "^0.23.1"
langchain-anthropic = "^0.1.6"
typer = "^0.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
